Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@1105 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ bindiff/src/BinDiff.java
@@ -41,7 +41,7 @@ public class BinDiff
 	private int cinsert;
 	private int posinsert = -1;
 	private TellStream fileinsert;
-	private int lastmark;
+	private long lastmark;
 
     public BinDiff(RandomAccessFile f1, RandomAccessFile f2, int cMinMatch, int cMaxSearch)
     {
